# docs/swagger.yaml
basePath: /
definitions:
  api.shortenRequest:
    properties:
      url:
        type: string
    type: object
  api.shortenResponse:
    properties:
      short_url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a high-level URL shortener written in Go.
  title: URL Shortener API
  version: "1.0"
paths:
  /{code}:
    get:
      description: Redirects from a short URL code to the original URL
      parameters:
      - description: Short code
        in: path
        name: code
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "302":
          description: Redirect
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Redirect to original URL
      tags:
      - main
  /analytics/{shortCode}:
    get:
      description: Returns daily click counts for a short URL
      parameters:
      - description: Short Code
        in: path
        name: shortCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get daily click counts
      tags:
      - main
  /analytics/top:
    get:
      description: Returns top short codes with highest click counts
      parameters:
      - default: 10
        description: Number of top results to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get top most clicked URLs
      tags:
      - main
  /shorten:
    post:
      consumes:
      - application/json
      description: Takes a long URL and returns a shortened version
      parameters:
      - description: URL to shorten
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.shortenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.shortenResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Shorten a URL
      tags:
      - main
swagger: "2.0"